requirements.test.txt
pytest
psycopg2-binary
sqlalchemy
testcontainers
docker
flask
flask-migrate
flask-sqlalchemy

